gn: add gn recipe from meta-arm to offer gn-native
With OpenHarmony and Matter we have two projects using GN as their build
system. To support the gn.bbclass we have in Oniro we need a gn-native
recipe as well to support building these projects.

While progress is being made to upstream gn-native to meta-oe we keep a
copy here to have a working build. The alternative would be to add two
extra layers to our build for just one recipe.

This file need to be removed once upstreamed into meta-oe.

The original source of this file is here:
https://git.yoctoproject.org/meta-arm/tree/meta-arm/recipes-devtools/gn/gn_git.bb

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>