meta-oniro-core: add and enable oniro-grub-bootconf
The oniro-grub-bootconf recipe provides the virtual-grub-bootconf
package which in turn contains a stub grub.cfg and empty grubenv file.

The change is spread around a considerable number of areas. Specifically
the kick-start file needs to explicitly reference the grub.cfg file which
must be placed in the wic/ directory. Since we need the same file in
grub/files directory, so that it can be installed into the image,
a symbolic link is employed.

The distribution configuration file, oniro-linux.conf sets
IMAGE_EFI_BOOT_FILES so that the grubenv file is installed into the boot
partition. This is required since grub cannot create the file by itself,
it can only edit the file, in-place, if it is present there already.

The actual configuration file for GRUB is a stub. Appropriate file with
some non-trivial logic will be proposed separately. The stub file boots
the first slot all the time. The kernel is loaded from the rootfs.

This test-boots fine in x86 and x86-64 qemu builds.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>