oniro-linux.conf: Force musl as TCLIBC
Upstream moved the default definition of TCLIBC as part of:

commit 57aa60ef6422568b425b6ccc4451567efc578469
Author: Martin Jansa <Martin.Jansa@gmail.com>
Date:   Wed Sep 16 15:28:52 2020 +0200
    bitbake.conf: use ${TCMODE}-${TCLIBC} directory for CACHE

This moved TCLIBC from defaultsetup.conf to bitbake.conf but the problem
is that it did it before the place were bitbake.conf includes the distro
configuration changing its behaviour. With this change, a ?= in the
distro configuration will be ignored as the first one is already
defined in bitbake.conf.

Because of the above, we ended up with defaulting to glibc. We revert
this by forcing the value of TCLIBC accordingly. This will have a side
effect that shell environment variable will be ignored. So doing that
through BB_ENV_EXTRAWHITE will be broken from now on. If we want that
ability in the future, we can try to address.

Fixes: https://booting.oniroproject.org/distro/oniro/-/issues/252

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>