sysota: Allow network in do_compile
sysota is a go project using go modules. Due to how this support is
currently implemented in the core, the recipe will download the
dependent go packages during compile.

On the other hand, lately, the core also defaulted to a restricted
network environment in the compile task breaking (but for good reasons)
the compilation of sysota. This situation was done as a forcing function
for the community to react and improve the current implementation of go
mods in the core. For now, we workaround it as this doesn't change the
current limitations/behaviour while we investigate better options in the
core.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>