sysota: disable shared linking
Shared linking has caused some problems on x86, where sysotactl would
crash on an apparently-stray pointer. This looks like a compiler bug.

Shared linking mode causes Go's standard library to be built into a
large .so file and all applications link to that with the dynamic
linker.

Given that sysota is the only Go application in our image and that is
has only one binary, this does not bring meaningful savings. In fact,
without shared linking, the overall image size is smaller as unused
parts of Go stdlib are removed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>