linux: enable CONFIG_SQUASH_ZSTD=y
This is useful for the outer squashfs used to deliver a RAUC update
bundle.  The inner squashfs uses stronger compression. The outer is now
using zstd (since a5437196). Having two layers of strong compression
makes no sense, as it cannot offer any improvements, and is much heavier
on CPU and memory.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>