ot-daemon: update to newer code base and switch to cmake
The main changes are the update to a newer upstream revision and the
switch to using the cmake build system of openthread. We also disabled
the fetch of external git submodules in third party to reduce the code
and involved licenses (which are not needed for the posix build).

Last but not least Thread 1.2 and spinel reset connection have been
enabled to work better with the meta-zephyr generated firmware images on
the radios.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>