create-spdx: fix error when symlink cannot be created
If IMAGE_NAME and IMAGE_LINK_NAME are the same an
exception is raised in image_combine_spdx because
it cannot create a symlink with the same exact
path of the original file. In that case there is
no point in creating a link in the first place.

Refactor image_combine_spdx to avoid code duplication

Upstream-Status: Accepted
http://git.openembedded.org/openembedded-core/commit/?id=88f76bba705007a7fcda506c79f743060d8f2e52

Signed-off-by: Davide Gardenal <davide.gardenal@huawei.com>