oniro-image: switch image compression from bz2 to gz
This makes is faster to build images, easier to handle them, e.g.
Raspberry Pi imager transparently supports .gz but not .bz2 and makes
virtually no difference to the size of the image.

Remember that the actual image is already compressed, as it is stored
inside the filesystem as a squashfs object.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>