.gitlab-ci.yml: publish RAUC bundle to HawkBit
The HawkBit instance needs to be configured externally and reachable
from the runner performing the publishing step. The location of the
instance must be provided through environment variables compatible with
hawkbitctl, namely HAWKBIT_{URL,USERNAME,PASSWORD}. HAWKBIT_TENANT is
optional and untested.

The bundle is added as an artifact to a new software module named after
the bundle, using bundle timestamp as the version. The module is then
placed in a distribution named after the image (not the bundle), with
the exact same version.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>