Submit jobs to squad instead of to LAVA directly
This is a first patch to introduce squad as a middleman in handling
LAVA job. SQUAD provides a DB to store test results in a scalable manner,
while also providing a good comparison engine for quick discovery of
regressions. Any visualization tool will make good use of SQUAD db and API.

Signed-off-by: Stevan Radaković <stevan.radakovic@linaro.org>