.gitlab-ci.yml: share rules among build, bundle, publish jobs
When looking at pipeline [1] we have realized that the newly introduced
bundle build and publishing jobs pose a problem for unrelated updates
of the tree, as depending on which exact files are modified by the
incoming pull request, the build jobs may not spawn, causing a
dependency error on the bundle and publishing jobs.

When initially introduced, it was overlooked that the jobs both have a
dependency on one another (build->bundle->publish) *and* come with a
distinct set of rules that govern if a specific job is created in a
given pipeline.

To resolve the problem move the rules to a new dot-job, called
.workspace-rules, and inject the dependency on .workspace-rules to all
the jobs that share the dependency chain.

[1] https://gitlab.eclipse.org/eclipse/oniro-core/oniro/-/pipelines/4285

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>