.oniro-ci,.gitlab-ci.yml: add optional clang jobs
Unqualified jobs, which rely on the default compiler, are gone.
Qualified jobs explicitly configure either clang or gcc and use the
appropriate suffix. Qualified jobs have needs/depends set to empty to
both avoid waiting for previous stages to start and to avoid pulling in
artifacts from the previous stage.

Qualified jobs inhabit two new lanes: build-gcc and build-clang, to
visually break the already long column of "build" jobs.

Job ordering within a lane is not exactly matching, gcc-clang. This is
noticeable with the jobs using the "extra" suffix. The reason for this
is unclear.

Some jobs are marked as broken as they do not build with clang. Allow
them to fail and make them start manually until more patches arrive.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>