skopeo: Add oniro-specific configuration
Skopeo is the component that podman uses to comunicate with container
registries so we configure skopeo for the OS requirements of the
project.

The most notable change is that we use the application partition for the
root's graphroot. This is where the data (including images) is stored
for the root user.

We also explicitly set the storage driver and runtime directory.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>