oniro-ci: Add variable to specify oe-core path or branch
This allows overriding of oe-core to point to different branch/repo for
testing updated which are not in stable repo yet or testing the repo
before bumping revision in the manifest.

The main purpose is testing oniro against <branch>-nut in scheduled
pipelines to avoid breakages.

To test oniro against different branch/repo of oe-core set pipeline
variables as follows:

	CI_ONIRO_OECORE_BRANCH: branch_to_be_tested_with
	CI_ONIRO_OECORE_REPO: url_of_the_repo

Example:

	CI_ONIRO_OECORE_BRANCH: stable/kirkstone-nut
	CI_ONIRO_OECORE_REPO: https://git.openembedded.org/openembedded-core-contrib

Signed-off-by: Pavel Zhukov <pavel.zhukov@huawei.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>