oniro.inc: Force bbappends parser behaviour to avoid surpises
BB_DANGLINGAPPENDS_WARNONLY controls how the parser behaves when
dangling bbappends are found. When not defined (or defined to 0/no), it
errors out and forces for these to be updated and handled accordingly.
We rely on this behaviour to force us to forward port changes or drop if
needed. The problem is that some layers might (in various ways) define
this to only warn in this case breaking our CI and project health
assumptions.

To avoid this, we force this in the inc included by all our distros.
Doing this in local.conf is problematic this would change behaviour by
just importing a layer.

Fixes: https://gitlab.eclipse.org/eclipse/oniro-core/oniro/-/issues/751

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>