linux-oniro-tweaks-all.inc: Define defconfig for qemuarm-efi based on multi_v7_defconfig
The multi_v7_defconfig file provides the initial support for this
machine but it also provides a good set of extra configs unrelated to
qemu. Some even generate build warnings due to some configs being
dependent on some that are not enabled. As an initial solution we pull
the defconfig in the layer and drop the problematic (warning generating)
configs. The plan is to drop it in the future for a config fragment for
qemu efi boot and reuse the existing BSP configuration from
yocto-kernel-cache of the equivalent qemuarm machine.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>
Signed-off-by: Luca Seritan <luca.seritan@huawei.com>