linux-oniro: Fix COMPATIBLE_MACHINE regex match
With the current regex expression a machine that is not part of the
compatible one would match if it is a substring of one of the machines
part of the regex expression.

For example "qemuarm64-efi" will match
"qemuarm|qemuarm64|qemux86|qemux86-64|qemuriscv64|qemuriscv32|qemu-generic-arm64".

Fix this by matching the start and the end of the string.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>
Signed-off-by: Luca Seritan <luca.seritan@huawei.com>