Align ONIRO_LAYERS_CONF_VERSION/ONIRO_REQUIRED_LAYERS_CONF_VERSION
We define the initial value for these variables as being 1. We also need
to set a default value for ONIRO_LAYERS_CONF_VERSION because the check
will not be triggered when ONIRO_LAYERS_CONF_VERSION is not defined
(builds before this check is introduced).

Finally, we drop POKY_BBLAYERS_CONF_VERSION as it is  an artefact from
when the project was using the poky distro.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>