meta-oniro-core: enable tpm as a distro feature
Enable TPM-related distro features. We're interested in TPM2, so
we need to enable both tpm and tpm2.

Also enable security dependencies. The dependency on meta-security
we have already had before, now add also tpm and tpm2 as we enable
the feature.

As a next step we need to enable related MACHINE_FEATURES for the
machines TPM (or software TPM) is available.

Signed-off-by: Marta Rybczynska <marta.rybczynska@huawei.com>