rauc-hawkbit-updater: fix typo in is-provisioned.conf
The correct syntax is ConditionFoo=, not ConditionalFoo=. This was
easily missed because invalid syntax is not fatal, and the service will
work correctly regardless of presence or absence of this configuration
element.

https://www.freedesktop.org/software/systemd/man/systemd.unit.html#ConditionPathExists=

Thanks to Kristijonas Čiulčinskas <kristijonas.ciulcinskas@huawei.com>
Grzegorz Gwóźdź <grzegorz.gwozdz@huawei.com> for reporting this.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>