modbus-testing: Add initial recipe for libmodbus examples
The libmodbus project does not offer examples of library usage in their
distribution. To make the life easier for Oniro integrators and
developers we provide a basic set of examples to get a head-start when
developing an application using libmodbus.

The scope for these examples is Oniro, so the recipe here is not
targeted at staging (I have no plans to upstream this), but at the
meta-oniro-core layer.

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>