python3-checkbox-ng: Integrate 1.18.1
The four patches fix compatibility with current version of jinja2 and
have been sent upstream.
The separate checkbox-ng.service binary package provides checkbox-as-a-service for certain
remote testing and automated testing scenarios. Checkbox uses this
facility to resume from restart/suspend/crash and carry on testing.
This service is packaged separately since on Ubuntu it was not installed by default.

Signed-off-by: Le Van Quach <levan.quach@kalpa.it>