matter: Switch to 1.0 release preparation branch and bump SRCREV
The Matter project finally created a v1.0 branch which they use to
finalize the release. This patch switches to this branches and bumps
the SRCREV accordingly.

In addition it adds some extra fixes needed for the update. A change in
the door-lock example name the addition of the editline third_party lib
and a workaround for a broken symlink in the pigweed files.

https://gitlab.eclipse.org/eclipse/oniro-core/oniro/-/issues/35

Signed-off-by: Stefan Schmidt <stefan.schmidt@huawei.com>